added unit_test for to_string and to_float conversion

+++ src/unit_tests/variant/variant_conv_to_int16.cpp
@@ -270,18 +270,18 @@ TEST_CASE("variant::to_int16() - from int8_t", "[variant]")
         variant var = int8_t(50);
         REQUIRE(var.can_convert<int16_t>() == true);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == 50);
+        CHECK(var.to_int16(&ok) == int16_t(50));
         CHECK(ok == true);
 
         CHECK(var.convert(type::get<int16_t>()) == true);
-        CHECK(var.get_value<int16_t>() == 50);
+        CHECK(var.get_value<int16_t>() == int16_t(50));
     }
 
     SECTION("valid conversion negative")
     {
         variant var = int8_t(-60);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == int8_t(-60));
+        CHECK(var.to_int16(&ok) == int16_t(-60));
         CHECK(ok == true);
         CHECK(var.convert(type::get<int16_t>()) == true);
     }
@@ -322,18 +322,18 @@ TEST_CASE("variant::to_int16() - from int32_t", "[variant]")
         variant var = int32_t(32760);
         REQUIRE(var.can_convert<int16_t>() == true);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == int32_t(32760));
+        CHECK(var.to_int16(&ok) == int16_t(32760));
         CHECK(ok == true);
 
         CHECK(var.convert(type::get<int16_t>()) == true);
-        CHECK(var.get_value<int16_t>() == int32_t(32760));
+        CHECK(var.get_value<int16_t>() == int16_t(32760));
     }
 
     SECTION("valid conversion negative")
     {
         variant var = int32_t(-32760);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == int32_t(-32760));
+        CHECK(var.to_int16(&ok) == int16_t(-32760));
         CHECK(ok == true);
         CHECK(var.convert(type::get<int16_t>()) == true);
     }
@@ -366,18 +366,18 @@ TEST_CASE("variant::to_int16() - from int64_t", "[variant]")
         variant var = int64_t(32760);
         REQUIRE(var.can_convert<int16_t>() == true);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == int64_t(32760));
+        CHECK(var.to_int16(&ok) == int16_t(32760));
         CHECK(ok == true);
 
         CHECK(var.convert(type::get<int16_t>()) == true);
-        CHECK(var.get_value<int16_t>() == int64_t(32760));
+        CHECK(var.get_value<int16_t>() == int16_t(32760));
     }
 
     SECTION("valid conversion negative")
     {
         variant var = int64_t(-32760);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == int64_t(-32760));
+        CHECK(var.to_int16(&ok) == int16_t(-32760));
         CHECK(ok == true);
         CHECK(var.convert(type::get<int16_t>()) == true);
     }
@@ -408,13 +408,13 @@ TEST_CASE("variant::to_int16() - from uint8_t", "[variant]")
     SECTION("valid conversion positive")
     {
         variant var = uint8_t(50);
-        REQUIRE(var.can_convert<uint8_t>() == true);
+        REQUIRE(var.can_convert<int16_t>() == true);
         bool ok = false;
         CHECK(var.to_int16(&ok) == 50);
         CHECK(ok == true);
 
-        CHECK(var.convert(type::get<uint8_t>()) == true);
-        CHECK(var.get_value<uint8_t>() == 50);
+        CHECK(var.convert(type::get<int16_t>()) == true);
+        CHECK(var.get_value<int16_t>() == int16_t(50));
     }
 }
 
@@ -427,11 +427,11 @@ TEST_CASE("variant::to_int16() - from uint16_t", "[variant]")
         variant var = uint16_t(32760);
         REQUIRE(var.can_convert<uint16_t>() == true);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == uint16_t(32760));
+        CHECK(var.to_int16(&ok) == int16_t(32760));
         CHECK(ok == true);
 
-        CHECK(var.convert(type::get<uint16_t>()) == true);
-        CHECK(var.get_value<uint16_t>() == uint16_t(32760));
+        CHECK(var.convert(type::get<int16_t>()) == true);
+        CHECK(var.get_value<int16_t>() == int16_t(32760));
     }
 
     SECTION("too big")
@@ -454,11 +454,11 @@ TEST_CASE("variant::to_int16() - from uint32_t", "[variant]")
         variant var = uint32_t(32760);
         REQUIRE(var.can_convert<uint32_t>() == true);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == uint32_t(32760));
+        CHECK(var.to_int16(&ok) == int16_t(32760));
         CHECK(ok == true);
 
-        CHECK(var.convert(type::get<uint32_t>()) == true);
-        CHECK(var.get_value<uint32_t>() == uint32_t(32760));
+        CHECK(var.convert(type::get<int16_t>()) == true);
+        CHECK(var.get_value<int16_t>() == int16_t(32760));
     }
 
     SECTION("too big")
@@ -481,11 +481,11 @@ TEST_CASE("variant::to_int16() - from uint64_t", "[variant]")
         variant var = uint64_t(32760);
         REQUIRE(var.can_convert<uint64_t>() == true);
         bool ok = false;
-        CHECK(var.to_int16(&ok) == uint64_t(32760));
+        CHECK(var.to_int16(&ok) == int16_t(32760));
         CHECK(ok == true);
 
-        CHECK(var.convert(type::get<uint64_t>()) == true);
-        CHECK(var.get_value<uint64_t>() == uint64_t(32760));
+        CHECK(var.convert(type::get<int16_t>()) == true);
+        CHECK(var.get_value<int16_t>() == int16_t(32760));
     }
 
     SECTION("too big")
